pyproject.toml
[project]
name = "conversing-in-semantic-id"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "datasets>=2.14.0",
    "nltk>=3.9.1",
    "gensim>=4.3.0",
    "numpy>=1.24.0",
    "vector-quantize-pytorch>=1.23.2",
]

[dependency-groups]
dev = [
    "ruff>=0.12.11",
]

[tool.ruff]
# Set the maximum line length to 88.
line-length = 88
# Assume Python 3.12
target-version = "py312"

[tool.ruff.lint]
# Enable standard linting rules
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings  
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
]

[tool.ruff.lint.per-file-ignores]
# Tests can use magic values, assertions, and imports
"**/tests/**/*" = ["PLR2004", "S101", "TID252"]

[tool.ruff.lint.isort]
known-first-party = ["conversing_in_semantic_id"]

[tool.ruff.format]
# Use single quotes for strings.
quote-style = "double"
# Indent with spaces, rather than tabs.
indent-style = "space"
# Respect magic trailing commas.
skip-magic-trailing-comma = false
# Automatically detect the appropriate line ending.
line-ending = "auto"
